refactor dimensions so const-expr can be used
Dimensions keep the ast-statements as start and end which
eventually will be replaced by a LiteralInteger if it can be
evaluated.

this allows to collect types and constants in the visit-phase
and following that, replace the dimensions that consist of
complex expressoins (e.g. "a + b") with a literal-integer
once all constants are known.